+++ R/encounter.R
@@ -1,4 +1,4 @@
-encounter <- function(object,include=NULL,exclude=NULL,...)
+encounter <- function(object,include=NULL,exclude=NULL,debias=FALSE,...)
 {
   UD <- object
   check.projections(UD)
@@ -10,27 +10,49 @@ encounter <- function(object,include=NULL,exclude=NULL,...)
   axes <- UD[[1]]$axes
   AXES <- length(axes)
 
-  # Gaussian approximation
   CTMM <- lapply(UD,function(U){U@CTMM})
-  CTMM <- encounter.ctmm(CTMM,include=include,exclude=exclude)
+
+  # Gaussian approximation
+  STUFF <- encounter.ctmm(CTMM,include=include,exclude=exclude,debias=debias)
+  CTMM <- STUFF$CTMM
+  BIAS <- STUFF$BIAS # individual biases
+  bias <- STUFF$bias # pairwise biases
+
   DOF.area <- rep(DOF.area(CTMM),AXES)
   names(DOF.area) <- axes
 
+  # bias correction (precision)
+  for(i in 1:length(object))
+  {
+    UD[[i]]$PMF <- UD[[i]]$PDF * prod(UD[[i]]$dr)
+    if(debias) { UD[[i]]$PMF <- debias.volume(UD[[i]]$PMF,1/BIAS[i])$PMF }
+  }
+
   DIM <- dim(UD[[1]]$PDF)
-  PDF <- matrix(0,DIM[1],DIM[2])
+  PMF <- matrix(0,DIM[1],DIM[2])
   for(i in 1:(length(UD)-1))
   {
     for(j in (i+1):length(UD))
-    { PDF <- PDF + include[i,j] * UD[[i]]$PDF * UD[[j]]$PDF }
+    {
+      PROD <- UD[[i]]$PMF * UD[[j]]$PMF
+      if(debias) # bias correction (variance)
+      {
+        gamma <- sum(PROD) # don't muck with weights
+        PROD <- PROD / gamma
+        PROD <- debias.volume(PROD,bias[i,j])$PMF
+        PROD <- PROD * gamma # don't muck with weights
+      }
+      PMF <- PMF + include[i,j] * PROD
+    }
   }
 
   dV <- prod(UD[[1]]$dr)
-  GAMMA <- sum(PDF) * dV # useful for relative encounter rates!
-  PDF <- PDF / GAMMA
+  GAMMA <- sum(PMF) # useful for relative encounter rates
+  PMF <- PMF / GAMMA
 
   object <- list()
-  object$PDF <- PDF
-  object$CDF <- pmf2cdf( PDF * dV )
+  object$PDF <- PMF / dV
+  object$CDF <- pmf2cdf( PMF )
   object$axes <- axes
   object$r <- UD[[1]]$r
   object$dr <- UD[[1]]$dr
@@ -60,14 +82,34 @@ encounter <- function(object,include=NULL,exclude=NULL,...)
 
 ################
 # Gaussian encounters
-encounter.ctmm <- function(CTMM,include=NULL,exclude=NULL,...)
+encounter.ctmm <- function(CTMM,include=NULL,exclude=NULL,debias=FALSE,...)
 {
-  #check.projections(CTMM)
+  # check.projections(CTMM)
 
   # Gaussian / cumulants
   axes <- CTMM[[1]]$axes
   AXES <- length(axes)
   isotropic <- all(sapply(CTMM,function(C){C$isotropic}))
+
+  # Wishart DOFs
+  DOF <- sapply(CTMM,DOF.wishart)
+  # precision matrices
+  P <- lapply(CTMM,function(M){PDsolve(M$sigma)})
+  # pairwise DOFs
+  dof <- matrix(0,length(DOF),length(DOF))
+  for(i in 1:length(DOF))
+  {
+    for(j in 1:length(DOF))
+    {
+      dof[i,j] <- tr(P[[i]]+P[[j]])^2 / ( tr(P[[i]])^2/DOF[i] + tr(P[[j]])^2/DOF[j] )
+    }
+  }
+  # clamp DOFs
+  DOF <- clamp(DOF,AXES+2,Inf)
+  dof <- clamp(dof,AXES+2,Inf)
+  # relative biases of inverse-Wishart
+  BIAS <- DOF/(DOF-AXES-1)
+  bias <- dof/(dof-AXES-1)
 
   fn <- function(CTMM)
   {
@@ -81,14 +123,22 @@ encounter.ctmm <- function(CTMM,include=NULL,exclude=NULL,...)
       {
         Pi <- PDsolve(CTMM[[i]]$sigma)
         Pj <- PDsolve(CTMM[[j]]$sigma)
-        prec <- Pi + Pj
-        sigma <- PDsolve(prec)
+        if(debias) # bias correction (precision)
+        {
+          Pi <- Pi / BIAS[i]
+          Pj <- Pj / BIAS[j]
+        }
+        Pij <- Pi + Pj
+        sigma <- PDsolve(Pij)
 
         mu <- sigma %*% (Pi %*% CTMM[[i]]$mu[1,] + Pj %*% CTMM[[j]]$mu[1,])
         mu <- c(mu)
 
         # intrinsic weight from multiplying two densities and renormalizing
-        include[i,j] <- include[i,j] / sqrt( det(CTMM[[i]]$sigma) * det(CTMM[[j]]$sigma) * det(prec) )
+        include[i,j] <- include[i,j] / sqrt( det(CTMM[[i]]$sigma) * det(CTMM[[j]]$sigma) * det(Pij) )
+
+        # the interplay between the bias correction steps and weighting has to be the same here as for the AKDEs
+        if(debias) { sigma <- sigma / bias[i,j] } # bias correction (variance)
 
         IN <- IN + include[i,j]
         M1 <- M1 + include[i,j] * mu
@@ -125,5 +175,5 @@ encounter.ctmm <- function(CTMM,include=NULL,exclude=NULL,...)
   info <- mean.info(CTMM)
 
   CTMM <- ctmm(mu=mu,sigma=sigma,COV.mu=COV.mu,COV=COV,axes=axes,isotropic=isotropic,info=info)
-  return(CTMM)
+  return(list(CTMM=CTMM,BIAS=BIAS,bias=bias))
 }